requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
boto3==1.26.129
botocore==1.29.129
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-extensions==3.2.1
django-model-utils==4.3.1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
openpyxl==3.0.10
packaging==23.1
Pillow==9.5.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.15
xlrd==2.0.1
pdfminer.six
psycopg2-binary==2.8.5